Fix dense_masked to use CachedDistAttention
The registry mapped dense_masked to RelativePositionalAttention
which doesn't accept dist_2d. Changed to CachedDistAttention which
does accept dist_2d, matching the training behavior where dist_2d
is pre-computed once and shared across all layers.

Also fixed dtype mismatch: dist_2d is float32 but needs to match
query's float16 dtype for the mask computation.

+++ benchmark_attn/scripts/benchmarks/benchmark_sweep.py
@@ -196,7 +196,7 @@ def build_closure(method_key, layer_count, seq_len, knn_neighbors,
     else:
         knn_idx = None  # Will be computed inside the closure.
 
-    dist_2d = torch.cdist(coords[..., 1:].float(), coords[..., 1:].float())
+    dist_2d = torch.cdist(coords[..., 1:].float(), coords[..., 1:].float()).to(dtype)
 
     if with_knn and registry_entry.needs_knn:
         def closure():
